Allow custom page title header tags
Problem: The Guides page layout adds the `title` attribute as an h1, so for the Derisking guide this creates a duplicate h1. To keep with the practice of the legacy de-risking guide (with the lift-and-shift mindset), the title should become an h2.

We considered a `custom_title` or `skip_page_title` attribute that would skip rendering the title. This assumes the author would put in a title manually. I saw that as a downside, and decided to go a different route.

This commit:

- Adds `page_title_tag` attribute that allows the author to set an arbitrary tag to use for the title (such as an h2).
- Adds the page's first tag as a class on the header
- Updates custom styles to render the derisking title's h2 more in line with the original
- Updates all non-intro pages in the state and federal guides