Using $BRANCH env variable for Cloud Pages
-- What
Cloud Pages does NOT set the NODE_ENV variable to production for _any_
of its builds, including the main site branch.

Instead, we can use a variable that they _do_ automatically set in
Pages builds: $BRANCH, which just corresponds to the branch name.

Now, we check for NODE_ENV=production (for testing locally) OR
BRANCH=main. Since I assume we won't ever be developing on the main
branch, this should be a workable solution.