eslint ignore the build directory and remove ext
-- What
We do not need to specify the extensions in the command line for the
eslint runner (if we need specific extensions only in the future, we
can add these into the config file where they belong).

Additionally, found a bad situation where eslint was reading from the
site build directory if a local build it present, which resulted in
tens of thousands of errors as a result of transpiled code from
external libraries. Have added config values to ignore the build directory.