Remove client-side redirects for content, de-risking, ux guides
Why?
Now that the guides are fully prepared for release, this performs step 2 of the migration

Results
This will "soft" launch the guides by making the new guide available on the web and indexable by web crawlers.

Considerations
Content for these guides will now be available in two places on the web: on the old guide domain and on the new guide domain. This isn't great for SEO, so we want to perform steps 3, 4, and 5 shortly after accepting this change.