Add the `html_block` rule to the markdown renderer
The goal for all links within the guides content is to evaluate them to see whether or not they leave the
".gov" domain and if they do, we want to make sure that the class list of the link has `usa-link--external`
applied so that the link gets the external link icon shown when it renders on the page.

Most of that work is handled by the `link_open` rendering rule which scans all of the markdown links and then
another portion is handled by `html_inline` for intances where the HTML is placed inside the markdown document.
The research for issue #348 discovered that when an HTML page is included using the 11ty `include` syntax that
neither `link_open` nor `html_inline` parse links within the included HTML file.

This commit adds a third renderer rule which scans all chunks of HTML to see if it includes any links that
leave the ".gov" domain  and then attempts to make sure the `usa-link--external` class is applied.